requirements.txt
asgiref==3.5.2
Django==4.1
django-ckeditor==6.5.1
django-debug-toolbar==3.6.0
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.2.0
sqlparse==0.4.2
django-environ==0.9.0
